backend/report-api/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*